refactor masker for RISEImage into free function
This is necessary for distance_explainer to be able to reuse it.
